backend/src/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
postgres = "^4.0"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
